Stop a previous instance of DEPNotify
If you're running this script from a jamf policy, you need to wait until jamf enrollment is complete before you run `jamf policy -event`. This could mean it takes a while for DEPNotify to appear, and the user might start doing things. One approach for PreStage Enrollment computers is to open DEPNotify with a general message until enrollment completes. So this code kills the previous instance of DEPNotify before the rest of the script continues and opens DEPNotify.