Updated preferences flushing logic
Instead of 'rm' ing the preferences file, use killall cfprefsd and "defaults delete" to flush every element out of the plist file.  This is in keeping with Apple's recommended practice since Mac OS X 10.9, when cfprefsd was introduced.  

Simply deleting the file does *not* flush the preferences out of memory, and may cause conflicts.